requirements.txt
accelerate==0.26.1
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diffusers==0.26.1
fastapi==0.109.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
psutil==5.9.8
pydantic==2.6.0
pydantic_core==2.16.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sniffio==1.3.0
starlette==0.35.1
sympy==1.12
tokenizers==0.15.1
torch==2.2.0
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
zipp==3.17.0
